+++ tests/test_pysight.py
@@ -1,6 +1,0 @@
-
-import pysight
-
-
-def test_main():
-    assert pysight  # use your library here